realtimeconsumer/requirements.txt
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pika==1.3.2
pandas==2.2.2